requirements.txt
adafruit-io == 1.1.0
dweepy == 0.2.0
pyserial == 3.1.1
